修复mybatis gen识别tinyint 创建属性类型为byte问题

+++ src/main/java/com/test/platform/hawkeye/service/Impl/AutoCaseInterfaceServiceImpl.java
@@ -29,7 +29,7 @@ public class AutoCaseInterfaceServiceImpl implements AutoCaseInterfaceService {
         for (AutoCaseInterface record :
                 autoCaseInterfaceList) {
             //逻辑删除
-            record.setIsDelete( (byte) 1 );
+            record.setIsDelete( 1 );
             AutoCaseInterfaceExample autoCaseInterfaceExampleIn = new AutoCaseInterfaceExample();
             autoCaseInterfaceExampleIn.createCriteria();
             autoCaseInterfaceExampleIn.createCriteria().andProjectIdEqualTo( projectId );
